Refactor some cool stuff
