+++ amplify/backend/function/bridgeXYZMutation/src/handlers/kycLinks.js
@@ -7,32 +7,10 @@ const { graphqlRequest } = require('../utils');
  */
 const { updateUser } = require('../graphql');
 
-let appSyncApiKey = 'da2-fakeApiId123456';
-let apiKey = 'da2-fakeApiId123456';
-let apiUrl = 'http://mocking:3000/bridgexyz';
-let graphqlURL = 'http://localhost:20002/graphql';
-
-const setEnvVariables = async () => {
-  const ENV = process.env.ENV;
-
-  if (ENV === 'qa' || ENV === 'prod') {
-    const { getParams } = require('/opt/nodejs/getParams');
-    [appSyncApiKey, apiKey, apiUrl, graphqlURL] = await getParams([
-      'appsyncApiKey',
-      'bridgeXYZApiKey',
-      'bridgeXYZApiUrl',
-      'graphqlUrl',
-    ]);
-  }
-};
-
-const kycLinksHandler = async (event) => {
-  try {
-    await setEnvVariables();
-  } catch (e) {
-    throw new Error('Unable to set environment variables. Reason:', e);
-  }
-
+const kycLinksHandler = async (
+  event,
+  { appSyncApiKey, apiKey, apiUrl, graphqlURL },
+) => {
   const checksummedWalletAddress = event.request.headers['x-wallet-address'];
   const { body, path } = event.arguments?.input || {};
