+++ amplify/backend/function/bridgeXYZMutation/src/index.js
@@ -1,7 +1,32 @@
 const { kycLinksHandler } = require('./handlers/kycLinks');
 const { putCustomerHandler } = require('./handlers/putCustomer');
 
+let appSyncApiKey = 'da2-fakeApiId123456';
+let apiKey = 'da2-fakeApiId123456';
+let apiUrl = 'http://mocking:3000/bridgexyz';
+let graphqlURL = 'http://localhost:20002/graphql';
+
+const setEnvVariables = async () => {
+  const ENV = process.env.ENV;
+
+  if (ENV === 'qa' || ENV === 'prod') {
+    const { getParams } = require('/opt/nodejs/getParams');
+    [appSyncApiKey, apiKey, apiUrl, graphqlURL] = await getParams([
+      'appsyncApiKey',
+      'bridgeXYZApiKey',
+      'bridgeXYZApiUrl',
+      'graphqlUrl',
+    ]);
+  }
+};
+
 exports.handler = async (event) => {
+  try {
+    await setEnvVariables();
+  } catch (e) {
+    throw new Error('Unable to set environment variables. Reason:', e);
+  }
+
   const { path } = event.arguments?.input || {};
 
   const handlers = {
@@ -15,5 +40,5 @@ exports.handler = async (event) => {
 
   const handler = handlers[path] || handlers.default;
 
-  return handler(event);
+  return handler(event, { appSyncApiKey, apiKey, apiUrl, graphqlURL });
 };
